iiod: Workaround cylic buffer race condition
There is an inherent race condition if a client creates a new cyclic buffer
shortly after destroying a previous. E.g. like

	iio_buffer_destroy(buf);
	buf = iio_device_create_buffer(dev, n, true);

In this case the two buffers each use their own communication channel which
are unordered to each other. E.g. the socket open might arrive before the
socket close on the host side, even though they were sent in the opposite
order on the client side. This race condition can cause an error being
reported back to the client, even though the code on the client side was
well formed and would work fine e.g. using the local backend.

To avoid this issue go to sleep for up to 50ms in intervals of 100us. This
should be enough time for the issue to resolve itself. If there actually is
contention on the buffer an error will eventually be returned in which case
the additional delay cause by the retires should not matter too much.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>